Reworked temporal_sync, finished preprocessing utils.
Gave up on the idea of maintaining separate fps levels during temporal sync and preprocessing. Instead, both vr and mocap data get upsampled to their fps's LCM and then the two are downsampled from there. This means the original vr input is not kept, but I don't think there's significant value in doing so (50 fps should be high enough that linear interpolation is a reasonable approximation).